+++ src/DateRangeInterface.php
@@ -18,28 +18,63 @@ use DateTimeInterface;
 interface DateRangeInterface
 {
     /**
-     * @return DateTimeInterface
-     */
-    public function getBegin(): DateTimeInterface;
-
-    /**
-     * @return DateTimeInterface
-     */
-    public function getEnd(): DateTimeInterface;
-
-    /**
+     * Returns whether the starting time is defined.
+     *
      * @return bool
      */
-    public function isFinite(): bool;
+    public function hasStartTime(): bool;
 
     /**
+     * Returns whether the ending time is defined.
+     *
+     * @return bool
+     */
+    public function hasEndTime(): bool;
+
+    /**
+     * Returns the starting time point.
+     *
+     * @return DateTimeInterface
+     */
+    public function getStartTime(): DateTimeInterface;
+
+    /**
+     * Returns the ending time point.
+     *
+     * @return DateTimeInterface
+     */
+    public function getEndTime(): DateTimeInterface;
+
+    /**
+     * Sets new starting time point.
+     *
+     * @param DateTimeInterface $time
+     *
+     * @return DateRangeInterface
+     */
+    public function setStartTime(DateTimeInterface $time): DateRangeInterface;
+
+    /**
+     * Sets new ending time point.
+     *
+     * @param DateTimeInterface $time
+     *
+     * @return DateRangeInterface
+     */
+    public function setEndTime(DateTimeInterface $time): DateRangeInterface;
+
+    /**
+     * Tells whether range is started at specific time.
+     *
      * @param DateTimeInterface $time
      *
      * @return bool
      */
-    public function isBeginAt(DateTimeInterface $time): bool;
+    public function isStartAt(DateTimeInterface $time): bool;
 
     /**
+     * Tells whether range is ended at specific time.
+     *
      * @param DateTimeInterface $time
      *
      * @return bool
@@ -47,26 +82,16 @@ interface DateRangeInterface
     public function isEndAt(DateTimeInterface $time): bool;
 
     /**
+     * Tells whether range is started at current time.
+     *
      * @return bool
      */
     public function isStarted(): bool;
 
     /**
+     * Tells whether range is ended at current time.
+     *
      * @return bool
      */
     public function isEnded(): bool;
-
-    /**
-     * @param DateTimeInterface $time
-     *
-     * @return DateRangeInterface
-     */
-    public function beginAt(DateTimeInterface $time): DateRangeInterface;
-
-    /**
-     * @param DateTimeInterface $time
-     *
-     * @return DateRangeInterface
-     */
-    public function endAt(DateTimeInterface $time): DateRangeInterface;
 }